requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.4
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
multidict==4.7.6
Pillow==7.2.0
pycparser==2.20
pylint==2.5.3
PyNaCl==1.3.0
python-dotenv==0.15.0
requests==2.24.0
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
